+++ src/service.rs
@@ -8,7 +8,7 @@ use std::ptr;
 use std::time::Duration;
 use std::{io, mem};
 
-use widestring::{NulError, WideCStr, WideCString, WideString};
+use widestring::{error::ContainsNul, WideCStr, WideCString, WideString};
 use windows_sys::{
     core::GUID,
     Win32::{
@@ -1740,7 +1740,7 @@ const MAX_QUERY_BUFFER_SIZE: usize = 8 * 1024;
 
 fn to_wide_slice(
     s: Option<impl AsRef<OsStr>>,
-) -> ::std::result::Result<Option<Vec<u16>>, NulError<u16>> {
+) -> ::std::result::Result<Option<Vec<u16>>, ContainsNul<u16>> {
     if let Some(s) = s {
         Ok(Some(
             WideCString::from_os_str(s).map(|s| s.into_vec_with_nul())?,
@@ -1781,7 +1781,7 @@ fn string_from_guid(guid: &GUID) -> String {
 
 pub(crate) fn to_wide(
     s: Option<impl AsRef<OsStr>>,
-) -> ::std::result::Result<Option<WideCString>, NulError<u16>> {
+) -> ::std::result::Result<Option<WideCString>, ContainsNul<u16>> {
     if let Some(s) = s {
         Ok(Some(WideCString::from_os_str(s)?))
     } else {
@@ -1790,7 +1790,7 @@ pub(crate) fn to_wide(
 }
 
 /// Escapes a given string, but also checks it does not contain any null bytes
-fn escape_wide(s: impl AsRef<OsStr>) -> ::std::result::Result<WideString, NulError<u16>> {
+fn escape_wide(s: impl AsRef<OsStr>) -> ::std::result::Result<WideString, ContainsNul<u16>> {
     let escaped = shell_escape::escape(Cow::Borrowed(s.as_ref()));
     let wide = WideCString::from_os_str(&escaped)?;
     Ok(wide.to_ustring())